FIX: Adding a section also adds the first group and command.
The section editor table is no longer fully disabled at program load.
Adding the first group and command to a new section makes it so that a group/command can be selected. Before this,  it was impossible to add the first group because a group was not available yet. Same for the first command.